modifying patches to apply policy
It used to be retrieving at most only the 5 last patches from the beginning of
the civil month when only the monthly stock file has been imported
(without daily patches). It means that if the last monthly stock file
does not match the current civil month, no patch (daily updates
from the previous months) is applied if the database is not already
up-to-date. It was particularly troublesome for first-time migration.