Corrected logic of Stock.importable ?
Previously, we say yes-importable if Stock is
same_as_current_stock_errored? or newer_than_current_completed_stock?

This is wrong in some edge-cases, such as:
"last stock imported is an old one but errored (expected yes, get no)"

This splitting of Stock.importable? in a set of cases should help with clarity.